🚀 Windows UTF Encoding
Currently fixing UTF encoding bugs, as well as removing extra MD files that don't need to be in the main repo. We've also added extra attributions to the attributions file. Template was old and out of date. If it's still required we'll have to fix it later on. Added colorama, and some extra cuda checks for windows. Qwen seems to like to argue wether 45.5 is bad for bits and bytes and so we'll see what happens.